pyproject.toml
[project]
name = "netcheckup"
version = "0.1.2"
description = "A simple internet health diagnostic tool"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [{name = "akhilsudhakaran"}]
dependencies = ["speedtest-cli"]

[project.scripts]
netcheckup = "netcheckup.__main__:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/AkhilSu/netcheckup"
Repository = "https://github.com/AkhilSu/netcheckup"
